backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
httpx==0.28.1
qdrant-client==1.15.1
pydantic==2.11.7
pydantic-settings==2.10.1
numpy==2.2.6
python-pptx==1.0.2
pypdf==6.0.0
pillow==11.3.0
pytesseract==0.3.13
